feat: benchmark script
